+++ arn/node.go
@@ -19,14 +19,13 @@ type Node struct {
 
 func (n *Node) match(result *[]*Node, parts []string, i int) {
 	lastPart := len(parts)-1 == i
-	// wip handle # /
 	switch {
 	case i > len(parts)-1:
 		*result = append(*result, n)
 		return
 
 	case n.txt == "#":
-		if len(parts[0]) > 0 && parts[0][0] != '$' {
+		if len(parts[0]) == 0 || parts[0][0] != '$' {
 			// https://docs.oasis-open.org/mqtt/mqtt/v5.0/os/mqtt-v5.0-os.html#_Toc3901246
 			*result = append(*result, n)
 		}