+++ router.go
@@ -7,25 +7,30 @@ import (
 	"sync"
 
 	"github.com/gregoryv/mq"
-	"github.com/gregoryv/tt/arn"
 )
 
 // newRouter returns a router for handling the given subscriptions.
 func newRouter() *router {
 	return &router{
-		rut: arn.NewTree(),
-		log: log.New(log.Writer(), "router ", log.Flags()),
+		log:     log.New(log.Writer(), "router ", log.Flags()),
+		filtSub: make(map[string][]*subscription),
 	}
 }
 
 type router struct {
-	m   sync.Mutex
-	rut *arn.Tree
 	log *log.Logger
+
+	// topic filter -> subscription
+	m       sync.RWMutex
+	filtSub map[string][]*subscription
 }
 
 func (r *router) String() string {
-	return plural(len(r.rut.Leafs()), "subscription")
+	var c int
+	for _, s := range r.filtSub {
+		c += len(s)
+	}
+	return plural(c, "subscription")
 }
 
 func plural(v int, word string) string {
@@ -40,12 +45,7 @@ func (r *router) AddSubscriptions(v ...*subscription) {
 	defer r.m.Unlock()
 	for _, s := range v {
 		for _, f := range s.filters {
-			n := r.rut.AddFilter(f)
-			if n.Value == nil {
-				n.Value = v
-			} else {
-				n.Value = append(n.Value.([]*subscription), s)
-			}
+			r.filtSub[f] = append(r.filtSub[f], s)
 		}
 	}
 }
@@ -59,24 +59,20 @@ func (r *router) removeFilters(sc *sclient, filters []string) {
 
 // Route routes mq.Publish packets by topic name.
 func (r *router) Route(ctx context.Context, p mq.Packet) error {
+	r.m.RLock()
+	defer r.m.RUnlock()
 	switch p := p.(type) {
 	case *mq.Publish:
-		// optimization opportunity by pooling a set of results
-		var result []*arn.Node
-		r.rut.Match(&result, p.TopicName())
-		for _, n := range result {
-			if n.Value == nil {
-				r.log.Println("node.Value is nil", n.Filter())
-				continue
-			}
-			for _, s := range n.Value.([]*subscription) {
-				for _, h := range s.handlers {
-					if err := h(ctx, p); err != nil {
-						r.log.Println("handle", p, err)
+		for filter, subscriptions := range r.filtSub {
+			if match(filter, p.TopicName()) {
+				for _, s := range subscriptions {
+					for _, h := range s.handlers {
+						h(ctx, p)
 					}
 				}
 			}
 		}
+
 	}
 	return ctx.Err()
 }